server/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
google-cloud-documentai
google-cloud-vertexai
